Move namespace tests to path
