+++ src/type/path.rs
@@ -140,3 +140,56 @@ impl TypePath {
 		}
 	}
 }
+
+#[cfg(test)]
+mod tests {
+	use super::*;
+
+	#[test]
+	fn namespace_ok() {
+		assert_eq!(
+			Namespace::new(vec!["hello"]),
+			Ok(Namespace {
+				segments: vec!["hello"]
+			})
+		);
+		assert_eq!(
+			Namespace::new(vec!["Hello", "World"]),
+			Ok(Namespace {
+				segments: vec!["Hello", "World"]
+			})
+		);
+		assert_eq!(Namespace::new(vec!["_"]), Ok(Namespace { segments: vec!["_"] }));
+	}
+
+	#[test]
+	fn namespace_err() {
+		assert_eq!(Namespace::new(vec![]), Err(NamespaceError::MissingSegments));
+		assert_eq!(
+			Namespace::new(vec![""]),
+			Err(NamespaceError::InvalidIdentifier { segment: 0 })
+		);
+		assert_eq!(
+			Namespace::new(vec!["1"]),
+			Err(NamespaceError::InvalidIdentifier { segment: 0 })
+		);
+		assert_eq!(
+			Namespace::new(vec!["Hello", ", World!"]),
+			Err(NamespaceError::InvalidIdentifier { segment: 1 })
+		);
+	}
+
+	#[test]
+	fn namespace_from_module_path() {
+		assert_eq!(
+			Namespace::from_module_path("hello::world"),
+			Ok(Namespace {
+				segments: vec!["hello", "world"]
+			})
+		);
+		assert_eq!(
+			Namespace::from_module_path("::world"),
+			Err(NamespaceError::InvalidIdentifier { segment: 0 })
+		);
+	}
+}
